feat: new admin note view for scores
