+++ components/admin-notes/AdminNoteView.tsx
@@ -114,11 +114,8 @@ export default function AdminNoteView({
         >
           <StickyNote className="h-3 w-3 text-white/70 hover:text-white" />
           {showNotification && notify && (
-            <div className="absolute -top-1 -right-1">
-              <span className="relative flex h-3 w-3">
-                <span className="absolute h-4/6 w-4/6 animate-ping rounded-full bg-warning duration-1500" />
-                <span className="absolute h-4/6 w-4/6 rounded-full bg-warning-foreground" />
-              </span>
+            <div className="absolute -top-1 -right-1 z-50">
+              <div className="h-2 w-2 rounded-full bg-yellow-600 dark:bg-yellow-400" />
             </div>
           )}
         </Button>
